move StartPage.tsx and Welcome.tsx to /pages-form
